Cleanup and small bug fixes
DPA protection points building:
 - Move the builder part in its own file.
 - Support custom builders.
 - Fix issues with case of file defined
 - add a test for the file-defined points
