+++ src/harness/reference_models/dpa/test_data/points_alameda.json
@@ -1,0 +1,2 @@
+{"type": "MultiPoint",
+ "coordinates": [[-122.291778430115, 37.7675355761059], [-122.2978976395834, 37.767880230306574], [-122.30401684905178, 37.76822488450725], [-122.31013605852017, 37.76856953870792], [-122.31574852164567, 37.770912354180176], [-122.31604769141556, 37.77538013743616], [-122.31061327735557, 37.77621137981417], [-122.30840279790188, 37.78038852286398], [-122.30228192218009, 37.78007484877446], [-122.29842993041576, 37.77773488524994], [-122.30067474404336, 37.77254427202493], [-122.29532961766411, 37.770588089809046], [-122.30225900502906, 37.77021719440227]]}
